trading_automation/requirements.txt
MetaTrader5==5.0.45
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.2
pytz==2024.1
ta==0.11.0
